Make authentication more testable
