+++ src/authentication/types.ts
@@ -2,6 +2,9 @@ export const authServerBaseUrlInjectSymbol = Symbol('auth-server-base-url');
 export const gitlabRootPasswordInjectSymbol = Symbol('gitlab-root-password');
 export const jwtOptionsInjectSymbol = Symbol('token-verify-options');
 
+export const subEmailClaimKey: 'https://sub_email' = 'https://sub_email';
+export const teamTokenClaimKey: 'https://team_token' = 'https://team_token';
+
 export interface AccessToken {
   iss: string;
   sub: string;
